refactor: 테스트를 도와주는 용도의 클래스들을 별도의 패키지에 분리
