+++ backend/src/test/java/edonymyeon/backend/member/MemberIntegrationTest.java
@@ -3,7 +3,7 @@ package edonymyeon.backend.member;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import edonymyeon.backend.IntegrationTest;
+import edonymyeon.backend.support.IntegrationFixture;
 import edonymyeon.backend.member.application.dto.response.MyPageResponse;
 import edonymyeon.backend.member.domain.Member;
 import edonymyeon.backend.post.application.dto.response.MyPostResponse;
@@ -17,7 +17,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
 @SuppressWarnings("NonAsciiCharacters")
-public class MemberIntegrationTest extends IntegrationTest {
+public class MemberIntegrationTest extends IntegrationFixture {
 
     @Test
     void 회원_정보_조회시_OK를_응답한다() {
