+++ backend/src/test/java/edonymyeon/backend/member/docs/MemberControllerDocsTest.java
@@ -18,7 +18,7 @@ import static org.springframework.restdocs.request.RequestDocumentation.pathPara
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import edonymyeon.backend.DocsTest;
+import edonymyeon.backend.support.DocsTest;
 import edonymyeon.backend.consumption.domain.Consumption;
 import edonymyeon.backend.consumption.repository.ConsumptionRepository;
 import edonymyeon.backend.member.application.dto.request.PurchaseConfirmRequest;
