+++ backend/src/test/java/edonymyeon/backend/post/ui/PostControllerTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import edonymyeon.backend.EdonymyeonTest;
+import edonymyeon.backend.support.IntegrationTest;
 import edonymyeon.backend.global.controlleradvice.dto.ExceptionResponse;
 import edonymyeon.backend.member.domain.Member;
 import edonymyeon.backend.post.ImageFileCleaner;
@@ -29,7 +29,7 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
 @SuppressWarnings("NonAsciiCharacters")
 @AutoConfigureMockMvc
-@EdonymyeonTest
+@IntegrationTest
 class PostControllerTest implements ImageFileCleaner {
 
     private static MockMultipartFile 이미지1;
