+++ backend/src/test/java/edonymyeon/backend/support/DocsTest.java
@@ -1,19 +1,15 @@
-package edonymyeon.backend;
+package edonymyeon.backend.support;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.web.servlet.MockMvc;
 
 @AutoConfigureMockMvc
 @AutoConfigureRestDocs
 @RequiredArgsConstructor
-@EdonymyeonTest
+@IntegrationTest
 public class DocsTest {
 
     protected final MockMvc mockMvc;
